Revert "Successful 'run' on local machine (#10)"
This reverts commit 2a1241f05add57e74fcb7703a09a7b7e21fe083e.
