--- Project/Frontend/src/App.js
@@ -1,60 +1,0 @@
-import React, { useState } from "react";
-import Button from "./components/Button";
-import LoginForm from "./components/LoginForm";
-import Signup_one from "./components/Signup-one";
-// import {BrowserRouter as Router, Link} from 'react-router-dom';
-
-const App = () => {
-  const adminUser = {
-    email: "pranav@chatur.com",
-    password: "pranav",
-  };
-
-  const [user, setUser] = useState({ name: "", email: "" });
-  const [error, setError] = useState("");
-
-  const Login = (details) => {
-    console.log(details);
-
-    if (
-      details.email === adminUser.email &&
-      details.password === adminUser.password
-    ) {
-      console.log("Logged IN");
-      setUser({
-        name: details.name,
-        email: details.email,
-      });
-    } else {
-      console.log("Details do not Match");
-      setError("Details don't Match Bitch!");
-    }
-  };
-
-  const Logout = () => {
-    console.log("User Logged Out!");
-    setUser({ name: "", email: "" });
-  };
-
-  return (
-    // <Router>
-    <div className="App">
-      {user.email != "" ? (
-        <div className="welcome">
-          <h2>
-            {" "}
-            Welcome, <span> {user.name} </span>
-          </h2>
-          <button onClick={Logout}>Logout</button>
-        </div>
-      ) : (
-        <div>
-          <LoginForm Login={Login} error={error} />
-        </div>
-      )}
-    </div>
-    // </Router>
-  );
-};
-
-export default App;
